requirements.txt
python-telegram-bot==21.7
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0